+++ packages/map/package.json
@@ -62,7 +62,6 @@
   },
   "typesVersions": {},
   "devDependencies": {
-    "solid-js": "^1.8.7",
-    "solid-transition-group": "^0.2.3"
+    "solid-js": "^1.8.7"
   }
 }
